A complete and proper preconditioner interface
Basically solves https://github.com/SciML/OrdinaryDiffEq.jl/issues/1551 sans any extra performance issues. The interface is simply that the user gives:

```julia
Pr, Pl = precs(W,du,u,p,t,newW,Plprev,Prprev,solverdata)
```

in the associated algorithm definitions. This gives a slight generalization over the Sundials interface so it should be enough, sans the `jcur` thing that is still an ongoing question. The setup phase is just a different dispatch on this function, i.e.

```julia
Pr, Pl = precs(W,du,u,p,t,::Nothing,::Nothing,::Nothing,solverdata)
```

which is rather clean. `solverdata` is for backwards compatibility: it's going to not be documented for a bit but allow for slapping things like `gamma` or `dt` in there, and adding to the struct won't be breaking like changing the call signature would be.

This PR is currently untested and only implements it for one algorithm, and so getting this to merge will essentially just require adding proper tests of using preconditioners which require updated Jacobians, like incomplete LU-factorizations, showing that it improves the convergence of GMRES. And setting up the rest of the algorithms to have this preconditioner interface as well, which is the simple but tedious part.